Separate Players, HandCard class from Player file
(cherry picked from commit 3ca26155665871b015f7b9ebf444a896dea1a108)
